\name{clearCoCoObjects}
\alias{clearCoCoObjects}

\title{ Clear CoCo object }

\description{ Clear CoCo objects. }

\usage{
clearCoCoObjects(coco.object = NULL, silent = FALSE, pos = .GlobalEnv, 
          printWarnings = FALSE) 

}

\arguments{
  \item{coco.object}{ See \code{\link[CoCoRaw]{endCoCo}}. }
  \item{pos}{ The envirinment of the CoCo object. }
  \item{silent}{ Logical. }
  \item{printWarnings}{ Logical. }
}

% \details{ }

% \value{ TRUE }

% \references{ Badsberg, J.H.: Xlisp+CoCo, Aalborg University, 1996. }

\author{ Jens Henrik Badsberg }

% \note{  }

% \section{ToDo}{ }

% \seealso{ }

% \examples{ }

\keyword{ interface }% at least one, from doc/KEYWORDS
\keyword{ multivariate }% __ONLY ONE__ keyword per line
