\name{CoCoObjects-internal}
\title{Internal CoCoObjects objects}
\alias{.--C--CoCoClass}
\alias{.--C--CoCoIdClass}
\alias{.--C--CoCoModelClass}
% \alias{.--C--CoCoTestClass}
\alias{.clear.coco.objects}
\alias{.end.temporary.object}
\alias{.First.lib}
\alias{.First.lib.CoCoObjects}
% \alias{.--M--dynamic.Graph:.GlobalEnv}
% \alias{.--M--dynamic.Graph.Old:.GlobalEnv}
% \alias{.--M--label:.GlobalEnv}
% \alias{.--M--labelOfTest:.GlobalEnv}
% \alias{.--M--modifyModel:.GlobalEnv}
% \alias{.--M--subModifyModel:.GlobalEnv}
% \alias{.--M--subModifyModel.Old:.GlobalEnv}
% \alias{.--M--testEdge:.GlobalEnv}
% \alias{.--M--testEdge.Old:.GlobalEnv}
% \alias{.--M--testLabel:.GlobalEnv}
% \alias{.--M--testWidth:.GlobalEnv}
% \alias{.--M--width:.GlobalEnv}
% \alias{.--M--widthOfTest:.GlobalEnv}
\alias{.new.coco}
\alias{.new.coco.model}
\alias{.object.of.model}
\alias{.object.of.thing}
\alias{.onAttach}
\alias{.onLoad}
\alias{.packageName}
\alias{.recover}
\alias{.recover.coco}
\alias{.recover.model}
\alias{.recover.reference}
\alias{.recover.search}
\alias{.return.key}
\alias{.return.model.number}
\alias{.return.model.of.object}
\alias{.return.object.model.number}
\alias{.return.reference}
\alias{.return.type}
\alias{.set.coco.value}
\alias{.SetSlotValue}
\alias{.sub.recover.coco}
\alias{.sub.recover.coco.model}
\description{Internal CoCoObjects objects.}
\details{These are not to be called by the user.}
\keyword{internal}
