\name{CoCoObjects-package}
\alias{CoCoObjects-package}
\docType{package}
\title{CoCoObjects}
\description{
The package provides classes for objects for CoCo-sessions 
and classes for objects for models in sessions.

This package contains a few functions of interest to the user.

Because of the development of the object system in R and due 
to the development of the CoCo-packages these classes are 
for CoCo-objects kept in this separate package.
}
\details{
   \tabular{ll}{
      Package: \tab CoCoObjects  \cr
      Type:    \tab Package      \cr
      Version: \tab 0.1.7.5      \cr
      Date:    \tab 2009-04-03   \cr
   }
}
\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
