\name{CoCoIdClass-class}
\docType{class}
\alias{CoCoIdClass-class}

\title{Class "CoCoIdClass"}

\description{ }

\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("CoCoIdClass", ...)}.
}

\section{Slots}{
  \describe{
    \item{\code{.reference}:}{Object of class \code{"numeric"},
			      see \code{\link{CoCoClass-class}}. }
    \item{\code{.key}:}{Object of class \code{"character"},
			see \code{\link{CoCoClass-class}}. }
    \item{\code{.type}:}{Object of class \code{"numeric"},
			 see \code{\link{CoCoClass-class}}. }
    \item{\code{.title}:}{Object of class \code{"character"},
			  see \code{\link{CoCoClass-class}}. }
  }
}

\section{Methods}{
No methods defined with class "CoCoIdClass" in the signature.
}

\references{ Badsberg, J.H.: Xlisp+CoCo, Aalborg University, 1996. }

\author{ Jens Henrik Badsberg }

% \note{  }

\seealso{
  \code{\link{CoCoClass-class}}, 
  % \code{\link{CoCoIdClass-class}}, 
  \code{\link{CoCoModelClass-class}}, 
  \code{\link[CoCoCore]{coco.init}}, 
  or \code{\link[dynamicGraph]{GraphLatticeProto-class}}.
}

% \examples{ }

\keyword{classes}
